fix: use fully qualified image names (#148)
Using an unqualified image name might yield different results on
different systems. The deployment actually relies on docker's default,
which is 'docker.io'. Other system might not have this default, and
some systems might even pause and interactively ask the user for a
decision.

Providing a fully qualified image name solves this issues and ensures
that the intended image is pulled.